Introduce Mapping abstraction. (#64)
Intended as a shared type for records and map values.
