+++ metafix/src/main/java/org/metafacture/metafix/Record.java
@@ -16,19 +16,10 @@
 
 package org.metafacture.metafix;
 
-import java.util.Collection;
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.function.BiConsumer;
-
 /**
- * Represents a metadata record, i.e., a mapping of fields and values.
+ * Represents a metadata record, i.e., a {@link Mapping} of fields and values.
  */
-public class Record {
-
-    private static final String EMPTY = "";
-
-    private final Map<String, Object> map = new LinkedHashMap<>();
+public class Record extends Mapping {
 
     private boolean reject;
 
@@ -70,112 +61,10 @@ public class Record {
         return reject;
     }
 
-    /**
-     * Checks whether this record contains the metadata field.
-     *
-     * @param field the field name
-     * @return true if this record contains the metadata field, false otherwise
-     */
-    public boolean containsField(final String field) {
-        return map.containsKey(field);
-    }
-
-    /**
-     * Checks whether this record is empty.
-     *
-     * @return true if this record is empty, false otherwise
-     */
-    public boolean isEmpty() {
-        return map.isEmpty();
-    }
-
-    /**
-     * Gets the number of field/value pairs in this record.
-     *
-     * @return the number of field/value pairs in this record
-     */
-    public int size() {
-        return map.size();
-    }
-
-    /**
-     * Adds a field/value pair to this record, provided it's not {@code null}.
-     *
-     * @param field the field name
-     * @param value the metadata value
-     */
-    public void put(final String field, final Object value) {
-        if (value != null) {
-            map.put(field, value);
-        }
-    }
-
-    /**
-     * {@link #put(String, Object) Replaces} a field/value pair in this record,
-     * provided the field name is already {@link #containsField(String) present}.
-     *
-     * @param field the field name
-     * @param value the metadata value
-     */
-    public void replace(final String field, final Object value) {
-        if (containsField(field)) {
-            put(field, value);
-        }
-    }
-
-    /**
-     * Retrieves the field value from this record.
-     *
-     * @param field the field name
-     * @return the metadata value
-     */
-    public Object get(final String field) {
-        return map.get(field);
-    }
-
-    /**
-     * Removes the given field/value pair from this record.
-     *
-     * @param field the field name
-     */
-    public void remove(final String field) {
-        map.remove(field);
-    }
-
-    /**
-     * Retains only the given field/value pairs in this record.
-     *
-     * @param fields the field names
-     */
-    public void retainFields(final Collection<String> fields) {
-        map.keySet().retainAll(fields);
-    }
-
-    /**
-     * Removes all field/value pairs from this record whose value is empty.
-     */
-    public void removeEmptyValues() {
-        map.values().removeIf(EMPTY::equals);
-    }
-
-    /**
-     * Iterates over all field/value pairs in this record.
-     *
-     * @param consumer the action to be performed for each field/value pair
-     */
-    public void forEach(final BiConsumer<String, Object> consumer) {
-        map.forEach(consumer);
-    }
-
     @Override
     public String toString() {
         // TODO: Improve string representation? Include reject status, etc.?
-        return map.toString();
-    }
-
-    // TODO: Replace map accesses with record operations!
-    public Map<String, Object> temporarilyGetMap() {
-        return map;
+        return super.toString();
     }
 
 }
