+++ metafix/src/test/java/org/metafacture/metafix/RecordTest.java
@@ -20,200 +20,15 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 public class RecordTest {
 
     private static final String FIELD = "field";
-    private static final String OTHER_FIELD = "other field";
     private static final String VALUE = "value";
     private static final String OTHER_VALUE = "other value";
 
     public RecordTest() {
-    }
-
-    @Test
-    public void shouldNotContainMissingField() {
-        final Record record = new Record();
-        Assertions.assertFalse(record.containsField(FIELD));
-    }
-
-    @Test
-    public void shouldContainExistingField() {
-        final Record record = new Record();
-        record.put(FIELD, VALUE);
-
-        Assertions.assertTrue(record.containsField(FIELD));
-    }
-
-    @Test
-    public void shouldNotContainNullValue() {
-        final Record record = new Record();
-        record.put(FIELD, null);
-
-        Assertions.assertFalse(record.containsField(FIELD));
-    }
-
-    @Test
-    public void shouldBeEmptyByDefault() {
-        final Record record = new Record();
-        Assertions.assertTrue(record.isEmpty());
-    }
-
-    @Test
-    public void shouldNotBeEmptyAfterAddingValue() {
-        final Record record = new Record();
-        record.put(FIELD, VALUE);
-
-        Assertions.assertFalse(record.isEmpty());
-    }
-
-    @Test
-    public void shouldNotAddNullValue() {
-        final Record record = new Record();
-        record.put(FIELD, null);
-
-        Assertions.assertTrue(record.isEmpty());
-    }
-
-    @Test
-    public void shouldGetSizeOfDefaultRecord() {
-        final Record record = new Record();
-        Assertions.assertEquals(0, record.size());
-    }
-
-    @Test
-    public void shouldGetSizeAfterAddingValues() {
-        final Record record = new Record();
-        record.put(FIELD, VALUE);
-        record.put(OTHER_FIELD, OTHER_VALUE);
-
-        Assertions.assertEquals(2, record.size());
-    }
-
-    @Test
-    public void shouldNotGetMissingField() {
-        final Record record = new Record();
-        Assertions.assertNull(record.get(FIELD));
-    }
-
-    @Test
-    public void shouldGetExistingField() {
-        final Record record = new Record();
-        record.put(FIELD, VALUE);
-
-        Assertions.assertEquals(VALUE, record.get(FIELD));
-    }
-
-    @Test
-    public void shouldNotReplaceMissingField() {
-        final Record record = new Record();
-        record.replace(FIELD, VALUE);
-
-        Assertions.assertNull(record.get(FIELD));
-        Assertions.assertFalse(record.containsField(FIELD));
-    }
-
-    @Test
-    public void shouldReplaceExistingField() {
-        final Record record = new Record();
-        record.put(FIELD, VALUE);
-        record.replace(FIELD, OTHER_VALUE);
-
-        Assertions.assertEquals(OTHER_VALUE, record.get(FIELD));
-    }
-
-    @Test
-    public void shouldNotReplaceExistingFieldWithNullValue() {
-        final Record record = new Record();
-        record.put(FIELD, VALUE);
-        record.replace(FIELD, null);
-
-        Assertions.assertEquals(VALUE, record.get(FIELD));
-    }
-
-    @Test
-    public void shouldRemoveMissingField() {
-        final Record record = new Record();
-        record.remove(FIELD);
-
-        Assertions.assertNull(record.get(FIELD));
-        Assertions.assertFalse(record.containsField(FIELD));
-    }
-
-    @Test
-    public void shouldRemoveExistingField() {
-        final Record record = new Record();
-        record.put(FIELD, VALUE);
-        record.remove(FIELD);
-
-        Assertions.assertNull(record.get(FIELD));
-        Assertions.assertFalse(record.containsField(FIELD));
-    }
-
-    @Test
-    public void shouldRetainFields() {
-        final Record record = new Record();
-        record.put(FIELD, VALUE);
-        record.put(OTHER_FIELD, OTHER_VALUE);
-
-        record.retainFields(Arrays.asList(FIELD));
-
-        Assertions.assertTrue(record.containsField(FIELD));
-        Assertions.assertFalse(record.containsField(OTHER_FIELD));
-    }
-
-    @Test
-    public void shouldRetainNoFields() {
-        final Record record = new Record();
-        record.put(FIELD, VALUE);
-
-        record.retainFields(Arrays.asList());
-
-        Assertions.assertTrue(record.isEmpty());
-    }
-
-    @Test
-    public void shouldNotRetainMissingFields() {
-        final Record record = new Record();
-        record.put(FIELD, VALUE);
-
-        record.retainFields(Arrays.asList(FIELD, OTHER_FIELD));
-
-        Assertions.assertTrue(record.containsField(FIELD));
-        Assertions.assertFalse(record.containsField(OTHER_FIELD));
-    }
-
-    @Test
-    public void shouldRemoveEmptyValues() {
-        final Record record = new Record();
-        record.put(FIELD, VALUE);
-        record.put(OTHER_FIELD, "");
-
-        record.removeEmptyValues();
-
-        Assertions.assertTrue(record.containsField(FIELD));
-        Assertions.assertFalse(record.containsField(OTHER_FIELD));
-    }
-
-    @Test
-    public void shouldIterateOverFieldValuePairs() {
-        final Record record = new Record();
-        record.put(FIELD, VALUE);
-        record.put(OTHER_FIELD, OTHER_VALUE);
-        record.put("empty field", "");
-        record.put("_special field", 1);
-
-        final List<String> fields = new ArrayList<>();
-        final List<Object> values = new ArrayList<>();
-        record.forEach((k, v) -> {
-            fields.add(k);
-            values.add(v);
-        });
-
-        Assertions.assertEquals(Arrays.asList(FIELD, OTHER_FIELD, "empty field", "_special field"), fields);
-        Assertions.assertEquals(Arrays.asList(VALUE, OTHER_VALUE, "", 1), values);
     }
 
     @Test
